2018-10-23 Commit

+++ foyoedu-provider-base/src/main/java/com/foyoedu/base/service/impl/BaseServiceImpl.java
@@ -1,11 +1,13 @@
-package com.foyoedu.service.impl;
+package com.foyoedu.base.service.impl;
 
-import com.foyoedu.dao.BaseDao;
-import com.foyoedu.service.BaseService;
-import com.foyoedu.utils.TypeConvert;
+import com.foyoedu.common.utils.SqlUtils;
+import com.foyoedu.base.dao.BaseDao;
+import com.foyoedu.base.service.BaseService;
+import com.foyoedu.common.utils.TypeConvert;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.lang.reflect.Field;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -33,26 +35,37 @@ public class BaseServiceImpl<E> implements BaseService<E> {
         return dao.deleteById(id, TBALENAME, PK_ID);
     }
 
-    public boolean delete(String filterCondition) {
+    public boolean delete(String filterCondition) throws Exception {
+        if (SqlUtils.sqlValidate(filterCondition)) {
+            throw new Exception("SQL参数中含有非法字符");
+        }
         return dao.delete(TBALENAME, filterCondition);
     }
 
     public List<E> list(Integer pageNo, Integer pageSize, String filterCondition, String sortCondition) throws Exception {
-        List<Map<String, Object>> maps = dao.findPage((pageNo-1)*pageSize, pageSize, TBALENAME, PK_ID, filterCondition, sortCondition);
+        if (SqlUtils.sqlValidate(filterCondition) || SqlUtils.sqlValidate(sortCondition)) {
+            throw new Exception("SQL参数中含有非法字符");
+        }
+
+        List<Map<String, Object>> maps = dao.findPage((pageNo - 1) * pageSize, pageSize, TBALENAME, PK_ID, filterCondition, sortCondition);
         TypeConvert<E> typeConvert = new TypeConvert<E>(clazz);
-        return typeConvert.MapToPojo(maps);
+        return typeConvert.MapToList(maps);
     }
 
-    public boolean update(String setColumns, String filterCondition) {
+    public boolean update(String setColumns, String filterCondition) throws Exception {
+        if (SqlUtils.sqlValidate(filterCondition)) {
+            throw new Exception("SQL参数中含有非法字符");
+        }
         return dao.update(TBALENAME, setColumns, filterCondition);
     }
 
-    public boolean add(E obj, String inserColumns, String values) throws Exception{
-        TypeConvert<E> typeConvert = new TypeConvert<E>();
-        Map<String, Object> map = typeConvert.PojoToMap(obj);
+    public boolean add(E obj, String inserColumns, String values) throws Exception {
+        if (SqlUtils.sqlValidate(inserColumns) || SqlUtils.sqlValidate(values)) {
+            throw new Exception("SQL参数中含有非法字符");
+        }
+        Map<String, Object> map = new HashMap<>();
         map.put("insertColumns", inserColumns);
         map.put("values", values);
-        map.put("pk_id", PK_ID);
         map.put("table", TBALENAME);
         boolean b = dao.add(map);
         Field field = obj.getClass().getDeclaredField(PK_ID);
@@ -61,7 +74,10 @@ public class BaseServiceImpl<E> implements BaseService<E> {
         return b;
     }
 
-    public Integer totalCount(String filterConditon) {
+    public Integer totalCount(String filterConditon) throws Exception {
+        if (SqlUtils.sqlValidate(filterConditon)) {
+            throw new Exception("SQL参数中含有非法字符");
+        }
         return dao.totalCount(TBALENAME, filterConditon);
     }
 }
